Initial foundation of the newer API.
Credits:
The functional part of addr.go is comming from @whyrusleeping and have been uglyfied by me (I'm not sure if it even still works) due to a removal in anacrolix/utp.
