+++ go.mod
@@ -1,0 +1,10 @@
+module github.com/libp2p/go-utp-transport
+
+go 1.13
+
+require (
+	github.com/libp2p/go-libp2p-core v0.2.3
+	github.com/multiformats/go-multiaddr v0.1.1
+	github.com/multiformats/go-multiaddr-fmt v0.1.0
+	github.com/multiformats/go-multiaddr-net v0.1.1
+)
