--- package.json
@@ -1,41 +1,0 @@
-{
-  "author": "whyrusleeping",
-  "bugs": {
-    "url": "https://github.com/libp2p/go-utp-transport"
-  },
-  "gx": {
-    "dvcsimport": "github.com/libp2p/go-utp-transport"
-  },
-  "gxDependencies": [
-    {
-      "author": "whyrusleeping",
-      "hash": "QmUAQaWbKxGCUTuoQVvvicbQNZ9APF5pDGWyAZSe93AtKH",
-      "name": "go-multiaddr",
-      "version": "1.1.2"
-    },
-    {
-      "author": "whyrusleeping",
-      "hash": "QmT6Cp31887FpAc25z25YHgpFJohZedrYLWPPspRtj1Brp",
-      "name": "go-multiaddr-net",
-      "version": "1.5.1"
-    },
-    {
-      "author": "whyrusleeping",
-      "hash": "QmQkdkvXE4oKXAcLZK5d7Zc6xvyukQc8WVjX7QvxDJ7hJj",
-      "name": "mafmt",
-      "version": "1.1.2"
-    },
-    {
-      "author": "whyrusleeping",
-      "hash": "QmWMia2fBVBesMerbtApQY7Tj2sgTaziveBACfCRUcv45f",
-      "name": "go-libp2p-transport",
-      "version": "2.2.1"
-    }
-  ],
-  "gxVersion": "0.4.0",
-  "language": "go",
-  "license": "MIT",
-  "name": "go-utp-transport",
-  "version": "1.6.0"
-}
-
